split header components into separate files
