+++ src/components/header/NameCard.js
@@ -1,0 +1,11 @@
+
+// Contains name and brief description
+export default function NameCard() {
+    return (
+        <div>
+            <h1>Param Tully</h1>
+            <h2>4th Year Computer Science Student at UBC</h2>
+            <p>based in Vancouver BC, Canada</p>
+        </div>
+    );
+}
